Subscribe e unsubscribe com read_all e write_all

+++ proj_24-25-p1_base/common/constants.h
@@ -4,12 +4,12 @@
 #define MAX_PIPE_PATH_LENGTH 40 // tamanho max do caminho do pipe
 #define MAX_STRING_SIZE 40
 #define MAX_NUMBER_SUB 10
-#define MAX_CONNECT_MESSAGE_SIZE ((size_t)121)  //tamanho max da mensagem de conexao 
+#define MAX_CONNECT_MESSAGE_SIZE ((size_t) 121)  //tamanho max da mensagem de conexao 
 #define MAX_CONNECT_RESPONSE_SIZE 3 //tamanho max da responsta da mensagem de conexao (op_code + char result + '\0')
-#define MAX_SUBSCRIBE_MESSAGE_SIZE 42 //tamanho max da mensagem de subscrição
-#define MAX_SUBSCRIBE_RESPONSE_SIZE 3 //tamanho max da resposta da mensagem de subscrição (op_code + char result + '\0)
-#define MAX_UNSUBSCRIBE_MESSAGE_SIZE 42 //tamanho max da mensagem de subscrição
-#define MAX_UNSUBSCRIBE_RESPONSE_SIZE 3 //tamanho max da resposta da mensagem de subscrição (op_code + char result + '\0)
+#define MAX_SUBSCRIBE_MESSAGE_SIZE ((size_t) 42) //tamanho max da mensagem de subscrição
+#define MAX_SUBSCRIBE_RESPONSE_SIZE ((size_t) 3) //tamanho max da resposta da mensagem de subscrição (op_code + char result + '\0)
+#define MAX_UNSUBSCRIBE_MESSAGE_SIZE ((size_t) 42) //tamanho max da mensagem de subscrição
+#define MAX_UNSUBSCRIBE_RESPONSE_SIZE ((size_t) 3) //tamanho max da resposta da mensagem de subscrição (op_code + char result + '\0)
 #define MAX_REQUEST_SIZE 42 //tamanho max para qualquer request
 #define MAX_KEY_SIZE 41 //tamanho max de um key
 #define MAX_DISCONECT_MESSAGE_SIZE ((size_t) 2) //tamanho max da messagem de desconexao (op_code) + '\0'
